+++ polkadot-appTests/Coinage/TestInstallation.swift
@@ -14,7 +14,7 @@ extension CoinageInstallationId {
 
 /// Lets the suites write an item as a plain number: it names that item in the test installation.
 extension CoinageKeyIndex: ExpressibleByIntegerLiteral {
-    public init(integerLiteral value: UInt32) {
+    public init(integerLiteral value: UInt64) {
         self.init(installation: .test, item: value)
     }
 }
